Fixed unintentional border glow in all Sapphire modules.
When I reworked the borders, I forgot that drawing on
layer 1 makes those things "lights" which causes them
to glow even when room lighting is turned down near 0%.

So all the Sapphire panel borders were glowing as if
outlined in neon.

I fixed this by moving the border rendering code
from `drawLayer` to `draw`. Now the borders are
part of the panel, not a glow above the panel.