Djinn panel rearrangement. Wider panel for main module.
The main/input module is now 12 HP wide instead of 9 HP.

Based on recent conversations with Omri and Fabio,
we want the feedback control as a single knob
on the input (main) module only.

So we needed the wider panel to make room for
global controls. Feedback will be the first global
control, and FREEZE will be another.

Added a REVERSE toggle group to each tap.