MultiTap: implemented global trigger group CLEAR.
Now pressing the clear button or receiving a trigger
on the clear input causes all tape loops to be cleared.

This is a cascade along the expander chain that travels
at the speed of one module per sample.
Therefore, each tap resets at a slightly different time,
on consecutive samples from the left to the right.