TapeLoop: prevent non-finite or high voltages from being recorded to tape.
My thinking is that feedback is inherently unstable.

The worst thing that can happen is for garbage to reach
the tape, because then it keeps recycling around and
the whole TapeLoop becomes corrupted.

Similar to how I have muted output for 1/4 second
any time a sample is bad (to avoid static), now I
mute the tape recorder for 1/4 second instead of
allowing even one sample of NAN/INF/huge voltages.