Echo: fixed #91 - equal time delay in forward and reverse directions.
The delay when playing the tape in the forward direction was correct.
But when in reverse mode, the echo was delayed by 10 seconds, which
is the maximum amount of buffer time.

The tape loop's total duration needs to be exactly twice the delay time,
so that the play head and the record head have the same duration
in either direction.

Also, I need to wrap around the tape by the delay time, not the buffer size.