Eliminate generated header file for Echo panel shift.
The panel generator script was also generating an include
file so that we would know how far to the right to move
the Echo panel's tape loop to the right to match spacing
with EchoTap.

But I can figure this out another way: by using the "_panel"
entry in sapphire_panel.cpp. Use the formula (x2-x1)/2.

Eliminated the generated header file.
I was going to rename it to "sapphire_echo_panel.hpp",
but getting rid of it is even better!