Fixed #84 - Echo expander button now creates correct module when there is a gap.
When a user clicks the button to create an expander, and there is
no module attached to the right side, search for the closest
module to the right within the HP width of EchoTap.

If that module exists and is EchoTap or EchoOut, create another
EchoTap to fill the gap and complete the expander chain.