Echo: fixes for TIME knob and selection of clock/seconds.
It is a lot simpler to default to factor=1 when in clock mode
and 1 second when in time mode. That way the numbers are
always the same.

Change the displayed text when operating a TIME knob
so that it says either "1.234 sec" or "X 1.234", depending
on the mode.

Have a two-level indicator for clock mode:
gray rectangle on the TIME knob when enabled
but there is no clock cable connected,
bright cyan rectangle when clock cable connected.

Limit operating time range to 0.1 seconds to 10 seconds.