Echo: implemented PAN control.
Now we have a simple left/right balance adjustment via
the PAN control. It currently only operates when the
number of channels is 2, but that is currently all we have anyway.

I may come back and provide other kinds of panning via a
right-click option on the PAN knob.
Or maybe even completely different stereo effects that
can be controlled with a parameter -1 <= x <= +1.