Echo: fixed #90 - red splash on overloaded tape loops.
When one of the tape loops gets overloaded, display
a red splash for 1 second.

Also clear the entire buffer as soon as corrupted
data would have been written.

Now I can customize the color of the splash to make
it look angrier in the overload case.

Also, before now I never splashed the leftmost Echo module.
Now that I do, and I'm splashing to indicate that the
tape loop is the problem, not the whole module, I can
now shift the left edge of the splash rectangle to the
right to include only the tape loop controls.
Doing this makes the splash look consistent across
all modules.