Echo: remove per-tap MIX control groups.
Based on a better understanding of Omri's signal flow idea,
I removed the MIX controls from each tap, leaving only the
global MIX control in EchoOut.

We don't want them because the output is not going to be just the final
tap; it is a mix of all tap outputs, whether we are in
serial mode or parllel mode (not yet implemented).

In future commits I will continue to rework the signal flow
to correct other misunderstandings I had.