Echo: finished correcting how final output is mixed.
Omri intended for the outputs of all taps to be mixed
in parallel to produce the final output, whether or
not the tap input mode is serial or parallel.

I added a new field to the expander bus called summedAudio.
It receives the sum of audio outputs from taps to the left.
Once the message reaches the EchoOut module, the sum
can be mixed and leveled to produce final output.