Refactoring DC reject filters into Sapphire base classes.
I use the same DC reject filter in Elastika, Nucleus, and Polynucleus.
There was duplicate code, and I now want to use the DC reject filter
in Echo also.

Instead of continuing a bad pattern, I am refactoring to make it
much easier to share the DC reject filter with new modules.
More precisely, the DC reject "Quantity", which doesn't actually
filter anything; it's just a value that gets saved/loaded via
JSON, and a slider in the context menus to change it.

Each consumer of the DC reject quantity will still have to figure
out what it means when the corner frequency is changed, in a
project-specific way.