Echo: work in progress - remove button for each tap
This is starting to work, but there is no undo/redo yet.

It also leaves a gap, but it should pull modules from the right
to fill the gap.

Also, when I close VCV Rack and re-open it, the module re-appears.

I will come back and look at this later when I'm ready to
study it more.