Echo: better reverse mode
The reverse mode was flawed because it was also reversing
what was written to the tape. I think the recorder should
always be doing the same thing, whether going forward or
in reverse. This preserves any interesting rhythmic content.

When you pop back out of reverse mode, everything should
be just like it would have been without the reverse.

I wonder if this is what should be sent to the next
level in the chain, in serial mode?