Echo: send forwards audio to next tap by default, even in reverse mode
I thought it might be too much chaos in serial mode
if reversed output is also sent to the next link in the chain.
But I don't want to prevent people from doing that.

So I made an option for changing the behavior of reverse output,
and default it to mix only. There is no UI control yet.

Creating a new pattern for representing enums in json.
It greatly reduces the verbosity!

Eliminated unnecessary local variable delayLineFeedback.
We especially didn't need to store all the channels;
a single float calculation per channel is fine.