Sapphire build process: no longer build all platforms locally.
It has been a long time since I had a compiler error
that showed up on a platform other than Linux.

I realized that every time I publish, the small chance
of catching a portability issue before pushing is small,
and GitHub Actions will catch it anyway, using a more
up-to-date cross-build process than I was using.

It wastes a lot of time to run the portability builds
twice before knowing the same result at the end, when
most of the time there are no portability problems anyway.

So I am no longer going to use my local cross-build
verification, especially because it has become stale
and I don't want to have to maintain it also.