Echo: in serial mode, do not flip both feedback to this tap and next tap.
I realized I was flipping the feedback loop in serial mode,
which adds chaos because it is also flipping to the next tap.
I wanted to have the option of flipping the feedback loop
because otherwe in parallel mode the FLP toggle would not do anything.

I still allow FLP to flip the feedback loop in parallel mode,
but in serial mode we always send forward audio to the feedback
loop, except in the the case of the first tap (Echo) where we
use audio from the end of the chain for the feedback.

In summary, FLP means "flip next" in serial mode,
or "flip feedback loop" in parallel mode.