Fixes for single-tap case: serial and parallel are identical.
When there is only a single tap (Echo module without EchoTap
to the right), and we are in Serial mode, do not use the
backward-path audio because there is nobody to send it to us.

A single-tap system has identical behavior whether we are in
serial mode or parallel mode. Therefore, it does not make sense
to offer a choice between S/P in that case. For single tap,
do not show S/P, and ignore any mouse clicks there.

Also cleaned up the code for the case where we want to tell
the difference between a fallback message and a real message
that was sent to us.