Echo: undo/redo for delete-tap is mostly working.
When you click the X at the bottom of a tap, it deletes the
tap and moves the remainder of the chain left to fill the gap.

You can undo the movement, and then undo the deletion, as
two separate steps. This is not ideal, so I will try to fix
this in a future commit.