Echo: removed unwanted feedback in serial mode for taps after the first.
The signal flow diagram had a mistake, which resulted in the
C++ code based on it to be wrong also. The mistake is that
in serial mode, only the first tap (EchoModule) should have feedback.
In that case it comes from the final module.

All taps after the first in serial mode should have no feedback at all.
The extra feedback was sending the loop into chaos, and not the good kind.