Echo: fixed #151 - do not create extraneous EchoOut on Windows.
Omri and I discovered that Rack on Windows causes the Echo module
to create another EchoOut module on every reload, even when there
is already an existing chain of modules.

Instead of playing games with guessing how many steps I should wait,
simply do not allow any Echo to create an EchoOut more than once.
Once the EchoOut is created, set a flag that prevents it from
happening again. The flag is persisted through JSON, so it will
work in the reload case.