Pop: fixed #134 - added tiny buttons: sync and pulse mode.
In Pop, you could always manually sync by right-clicking on the
menu and selecting "Sync polyphonic channels". But this is awkward.
I added a tiny green button to the right of the SYNC port that
does an immediate sync when you click on it.

Similarly, you could always change the output mode from triggers
to gates through the menu, but now it is easier to press the
new pulse mode button to the left of the PULSE output port.