Echo: fixed #161 - redo adding Echo without messing up action history.
To prevent destroying any future actions in the history stack,
when redoing insertion of the Echo module, suppress creating
an EchoOut module to the right if there are any redo actions
available. The VCV Rack SDK has a handy method just for this,
so I don't feel like I'm doing anything too weird.