Echo: Fixed #169 - Display CLOCK, FRZ, REV in panel preview.
In the screenshot of Echo in v2.5.9, both in the VCV Rack Library
and the module browser inside VCV Rack itself, 3 labels are missing:

- CLOCK/RATE label.
- REV/FLP label and surrounding hexagon.
- ENV/DCK label and surrounding rectangle.

In all cases, we have a 4-way state resulting from a pair of boolean toggles:
- Which of the two things is selected, e.g. REV or FLP.
- What color to draw the label: black (normal) or magenta (mouse hover).

Logically, given any group of 4 states that have mutually exclusive
visual representations, one of them must be visible while 3 are invisible.
This 1+3==4 should be treated as an invariant.
In user interface design lingo, they are conceptually like radio buttons.

The bug was that in preview mode, there is no module, nor are there any calls
to Widget::step(). Therefore, my code never got to the place where we would
have picked 1 of the 4 svg layers to become visible.

The module==nullptr / screenshot case violated the invariant by leaving
0 of the 4 visible.

I fixed this by allowing an optional `true` parameter to addLabelOverlay()
to make that label start out visible instead of hidden.