Fixed #172 - easier way to add Tricorder to right.
Thanks to Sapphire Echo, I know how to programmatically insert modules.
So I added a menu option to insert Tricorder to the right,
but only when it makes sense to do so. If the current module is
a vector sender, and there is not already a Tricorder immediately
to the right, allow adding a Tricorder.

I did it in the base classes SapphireWidget and SapphireModule,
so that every module that acts as a vector sender inherits the
new behavior.