Zoo: on reset, start at point on the Rossler attractor.
The Rossler preset worked correctly, but when I initialize Zoo,
it was not starting at the same point. It took a lot longer to
spiral out to the chaotic mode of operation.

So I copied the initial (x, y, z) location from the Rossler preset
back into the Zoo C++ source code.

I also fixed bugs with keeping json in sync with memory[0].