Zoo: added "speedFactor" to adjust how fast particles move.
The Qi-3D factory preset was moving too fast by default,
compared to the other attractors.

I added a speedFactor setting in the json that affects the
speed all the way down in the RK4 integrator, so that directions
are not modified (and therefore trajectory), but makes the
particle speed up or slow down by any factor from 0.01 to 100.

Qi-3D now has speedFactor = 0.1.

All the others have speedFactor = 1.0.