Zoo: design change - initialize to a blank state
I realized it was weird to initialize to the same state as one
of the presets (Rossler), when that preset had some strange
things like position and velocity morphing.

It is a lot easier to start with better defaults for everything,
and I also decided to use an ultra-safe default program:

   vx = 0
   vy = 0
   vz = 0

The point is locked at the initial location!

People can choose a factory default, then safe as their own
default settings if they want.