Fixed problem inserting Chaops expander on left.
The code for inserting an expander module on the left side
did not work when there was any other module already in
the landing zone for the new module. It caused the inserted
module to be separate from the original module, which
breaks the expander chain.

The way VCV Rack wants to do this is to shift everything
to the right, so I have to insert the expander at the
same location as the parent module.