Zoo: bug fix - need to reset 'fallback' positions each time loading JSON.
When the "fallback" option is absent from the module json,
we must overwrite our fallback value with NAN, instead of
keeping the existing value. Fallback is only present in factory
presets that need it. Those that need it should not have their
values leaking into others that don't need it.

Doing so causes incorrect parameter values that destablize
the modules into which the leakage occurs.