Vina: pre-settle the mesh before generating audio.
This is a general problem I need to solve for the Vina engine.
Anytime you want to generate audio, you need to make sure
the mesh starts out in a minimum potential energy state.
Otherwise, you will get an unwanted "pluck" at the beginning,
and it really doesn't sound very good.

This is the same problem that the Nucleus module had.
Nucleus and Vina engines both need to be "pre-settled".

I hacked this in the unit test with a simulated 10-second wait.
I may need to "bootstrap" the code so that we generate the
default state, then settle, then rewrite the ball positions
as another header file.

Then vina_vcv.cpp can include both header files and do the
wiring to make sure all VinaEngine objects start in a
minimum energy state.