Vina: starting to implement internal attack/decay logic.
Now I actually treat the gate input both like a trigger
to get the exciter started, and a gate to determine
the decay half-life. By keeping the half-life longer when
the gate is high, it acts a little like a sustain.
When the gate is low, I decrease the half-life for
a shorter release.