Vina: much improved string sound
Instead of using velocity to produce the output signal,
we now use position, which is the integral of velocity
and therefore reduces harsher high frequencies.

Then I use DC reject to ignore the fact that the position
does not center around zero.

After that, use a Gravy lowpass filter to reduce the
harshness of the upper frequencies.

I had to compensate the output level because of the filters.