Vina: Moving to a 1D string model.
I came to the conclusion that virtually all of the sound quality
problems are from having multiple modes of movement that are
difficult to tune to each other.

I realized that a 1D representation of a string, where all vibration
is longitudinal, is an interesting approach that eliminates
the tuning problem.

The result is a little bit like Karlplus-Strong, but I think I
can play all kinds of games with pre-filtering, post-filtering,
chorus, reverb, distortion, and maybe feedback for interesting
timbre and stereo effects.