Vina: more lateral refactoring
I'm getting ready to replace 16 (worst case) reverb units with 1.
But first I had to restructure so that the VinaEngine is now
a collection of VinaWires, all of which will feed through
a single reverb. This will improve performance, and will also
simplify a planned "standby" mode for each wire.