Vina: no more filter snapping
I was using a flag called isFirstSample to track when
it was necessary to lazy-initialize dcReject and audio filters.
The whole problem was that I was using raw positions as
inputs, and "snapping" to them as initial values got rid
of an initial unwanted impulse signal.

Now I realize it's so much simpler to capture the original
position values in the initialize function and use them forever.
Subtract and filter the difference. Then we can always
initialize filters relative to zero.

This commit includes work in progress for standby mode.