Vina: trigger detection belongs to the caller
Each VinaWire was doing its own gate-to-trigger conversion.
But I realized we might want to keep triggering while the gate
is held high for ratcheting. There are other possible creative
possibilities.

Refactored so that the topmost caller passes in both gate and trigger
as separate flags. This means we could conceivably pluck a string
while staying in the release decay state, etc.

This opens up the possibility of a separate TRIGGER gate
that, when connected to a cable, could override deducing
triggers from GATE.

Also started down the road of figuring out how to
dynamically allocate a wire to be plucked each time,
to implement Omri's idea of arpeggios decaying correctly
even in a "simple stereo" patch where L and R outputs
are both monophonic.