Vina: Omri's idea - dynamic wire assignment, first draft.
Allow mapping zero or more wires to a given channel.

Allow designating a specific wire as "active" for a given channel.
That means designating which wire receives gates and triggers
for that poly channel.

Sum contributions for all wires assigned to a channel
to produce the output audio frame.

This change allows playing arpeggios with long decays,
but comes at the expense of using up to 16 wires simultaneously
even in the monophonic case.