Refactored tiny button filenames to indicate their colors.
The way I named many of the tiny button SVG files did not make sense.
They were named after how Echo first used them, then appropriated
for other places that needed a button of the same color:
green, yellow, or red.

I decided to rename the files to indicate their color, which
is all that really matters about the SVG file itself.
Then it is explicit that a mute button should be "red", etc.

Added an inline function addTinyButtonFrames to make the
refactored code smaller and clearer in meaning.

Fixed usage in the modules Echo, Vina, and Pop.