Expander button bug fix - expander chain is not always available
I forgot that there are cases where no expander chain is
defined. I believe I encountered this developing Echo.

Now I encountered it testing expander buttons
on Windows when no audio device was yet selected in the patch.

Test: create Glee (or any chaos module) in an otherwise
empty patch. Do not create an AUDIO output module. If one
is already in the patch, do not select a device yet.

Click the "Insert Chaops" button repeatedly.
The bad behavior is if it keeps adding more Chaops with each
button press. Same bug happens for "Insert Tricorder".

I was only able to reproduce this in Windows.
For some reason, I could not make the bug happen in Linux.
I did not test on Mac.

This commit fixes the issue such that we can detect an adjacent
module to the left or right without using the expander chain.
Instead, we enumerate through the collection of all modules
and check for matching integer coordinates (hp, rail).